backend/pyproject.toml
[project]
name = "factwave"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core framework
    "crewai",
    "crewai-tools",
    "pydantic",
    "python-dotenv",
    "rich",
    "pyyaml",
    "fastapi",
    "uvicorn",
    # LLM providers
    "openai",
    "anthropic",
    "litellm",
    # Data processing
    "pandas",
    "numpy",
    "scipy",
    "scikit-learn",
    # HTTP requests
    "requests",
    "httpx",
    "aiohttp",
    # Academic tools
    "arxiv",
    "wikipedia-api",
    # Statistical tools
    "PublicDataReader",
    # OWID RAG system & ML
    "chromadb",
    "sentence-transformers",
    "rank-bm25",
    "torch",
    "transformers",
    "tokenizers",
    "tiktoken",
    # Language processing
    "nltk",
    "langchain",
    "langchain-core",
    "langchain-community",
    "langchain-openai",
    "langchain-text-splitters",
    "langsmith",
    # Web scraping & parsing
    "beautifulsoup4",
    "xmltodict",
    "playwright",
    "pdfplumber",
    "pypdf",
    # Database & storage
    "sqlalchemy",
    "psycopg2-binary",
    "lancedb",
    "qdrant-client",
    # Utilities
    "typer",
    "click",
    "tenacity",
    "joblib",
    "tqdm",
    "fire",
    "packaging",
    # Monitoring & observability
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "posthog",
    # Development
    "black",
    "isort",
    "pytest",
    "pytest-asyncio",
    "twscrape>=0.17.0",
    "websockets>=15.0.1",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
ko = [
  "konlpy>=0.6.0"
]
